Fixed assert when starting a turbor machine
Regression introduced yesterday in revision 12707, when we added the
di_halt_callback setting.

A turbor machine has both a Z80 and R800 CPU, we should not let both CPUs
create the di_halt_callback setting (because that triggers an assert: setting
names need to be unique). As a fix we moved the creation of the setting to the
MSXCPU class (just like this class already contains the shared 'cpu_trace'
setting).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12711 7512d703-071c-0410-b5c2-d2bdd895b7d1