Fixed 'Disk offline' message in AVT_DPF-550
Before this patch when executing the 'FILES' command in MSX-BASIC with the
AVT_DPF-550 extension without a disk in the drive, you got the wrong error
message 'File not found'. Now it returns the proper error message 'Disk
offline'.

The main cause for this error was that the AVT_DPF-550 uses a WD1770
controller, while we emulated a WD2793. One of the main differences between
these two controllers is that the WD1770 has no drive-ready input signal pin
(instead it has a motor-on/off output pin). 'Normally' when executing a FDC
command (e.g. reading a sector for the FILES command) the ready signal is
checked and if the drive is not ready, the command fails. But of course this is
not possible on WD1770. What the disk rom of the AVT_DPF-550 does instead is
execute the command and if after some time it doesn't get a response, it
assumes there was no disk in the drive.

After I fixed the above issue (I let the WD1770 ignore the ready signal), it
still didn't work. This was because of inaccurate timing emulation of errors on
sector read/write commands: When the WDxxxx receives a sector read/write
command, the first thing it does is locate the correct sector on the disk. This
takes some time because the disk has to rotate. If the sector is not found
within 5 revolutions, the WDxxxx gives an error. Before this patch this timing
aspect was emulated like this:
 - locate the correct sector in the track
 - if not found -> error
          found -> calculate distance between current and destination rotation
                   angle and wait for the appropriate amount of time
So successful read/writes already had the correct timing, but failures had not.
Now we emulate this timing closer to how the real hardware works: rotate to
next sector, if not yet correct sector repeat (for max 5 revolutions).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12732 7512d703-071c-0410-b5c2-d2bdd895b7d1