Allow to specify the initial content for RAM
We discovered that the built-in software of the TPC-310 machine relies on
specific initial content of the RAM. It doesn't work properly when RAM is
initialized to all 0xff: the pointer can't be moved using the cursor keys,
though it does work after a reset. This patch allows to specify the exact
initial content for the RAM. And when e.g. RAM is initialized to 0x00, the
software does work.

Some examples:
    <MemoryMapper id="Main RAM">
        <mem base="0x0000" size="0x10000"/>
        <size>1024</size>
        <initialContent encoding="hex">00ff</initialContent>
    </MemoryMapper>
    ...
    <RAM id="other ram">
        <mem base="0x4000" size="0x4000"/>
        <initialContent encoding="gz-base64">eNpj+M8wouEI9z4DAJBx/wE=</initialContent>
    </RAM>

So inside a RAM or MemoryMapper tag you can specify a pattern using the new
'initialContent' tag. This pattern will be repeated over the whole ram content.
For short patterns it's easiest to use the 'hex' encoding (1st example). For
longer patterns the gz-base64 encoding is more practical (e.g. the pattern in
the 2nd example expands to '128 x [00 ff], 128 x [ff 00]'). I've also added a
tool (Contrib/base64.c) to create the gz-base64 encoding for a pattern.

We can extend this to also allow to specify the initial content of VRAM, but
that's something for a follow-up patch (it may actually already work for more
devices than RAM and MemoryMapper).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12740 7512d703-071c-0410-b5c2-d2bdd895b7d1