Fixed bug '[3533002] Y8950 not being detected by MSX-Audio v1.3'
Apparently after a reset, the BUF_RDY bit should be set (when it's not masked).
Normally this bit indicates whether the adpcm unit is ready for the next byte
(e.g. when transfering data to/from memory or playing/recording samples from
CPU). In this patch I also set the BUF_RDY bit when the adpcm unit is not
performing any function.

The audio-detection routine in 'MSX-Audio BIOS v1.3' depends on this behavior.
It checks the value of the status register (after triggering a timer, but
without using the adpcm unit) and this check includes the BUF_RDY bit.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12746 7512d703-071c-0410-b5c2-d2bdd895b7d1