Fixed sector writes on Microsol_CDX-2 (possibly also other diskroms)
This was a regression introduced during the whole DMK refactoring. One of the
changes in this refactoring was more accurate timing for the DTRQ signal. After
this change, when the CPU doesn't supply the next byte in time, a 'LOST_DATA'
error is generated. And this error was now triggered by the write-sector
routine in then Microsol_CDX-2 disk rom.

But why does this error trigger? While clearly on the real hardware it works
just fine. This is what the write sector routine on the Microsol_CDX-2 looks
like:
	while (1) {
		if (command-finished) return
		if (ready-to-receive-next-byte) break
	}
	write-one-byte
	DELAY
	while (1) {
		if (command-finished) return
		if (ready-to-receive-next-byte) write-one-byte
	}

So after the first byte is written, there is an explicit delay in the code. I
have no idea why that delay is there. Functionally the code behaves the same if
only the 2nd while loop is present (and that's exactly how this routine is
coded in most other disk roms I've looked at). But this extra delay between the
1st and 2nd byte is causing the LOST_DATA error.

Ok, but why does this work on real hardware? When the WD2793 receives a write
sector command, it first locates the sector header. Next it activates the DTRQ
signal (so the CPU can send the first byte). Next it still has to wait for the
disk to rotate to the actual sector data. At that point the first byte will be
written and the DTRQ is activated again. For that point on DTRQ is activated at
more regular time intervals. (This is a simplified scenario, see code for more
details).

This patch implements the more accurate timing of the 1st DTRQ activation.
After this change the write sector routine on Microsol_CDX-2 works again.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12749 7512d703-071c-0410-b5c2-d2bdd895b7d1