Fixed crash on 'openmsx -diska ALESTE2.ZIP'
This ZIP file contained as first file 'ALESTE2.TXT', with file size 122 bytes.
(Currently openMSX always picks the first file from a ZIP). Later we try to
read the first sector (512 bytes) from this 'disk image'. That will read beyond
the end of the file and it triggered an assert.

If you try a similar thing with an non-(g)zipped file that is smaller than 512
bytes, it doesn't crash, (but of course such a disk image isn't very useful).

This patch converts the assert in the CompressedFileAdapter code into a check
for read-beyond-end-of-file. The regular file reading code already has such a
check.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12767 7512d703-071c-0410-b5c2-d2bdd895b7d1