Make loadstate code robust against unknown enum strings
In a (xml) savestate enums are serialized as strings instead of numerical
values (this gives more freedom to rearrange the numeric values of an enum
without breaking backwards compatibility with older savestates). Though the
enum-deserialization code didn't check for unknown strings (but normally such
strings should not occur). After this patch we now give a proper error message
on such invalid inputs.

The motivation for this patch was the recently committed revision 12801. That
commit removes the 'NEC_REPEAT_PULSE' value from the 'RemoteState' enum. In
theory this could break old savestates that use this enum value. I discussed
this with Sean, and it's _highly_unlikely_ there are existing savestates with
that value (no existing laserdisc code could ever reach that state). So instead
of adding backwards compatibility code we decided to really drop this old enum
value. And just in case we also made openMSX more robust against invalid
savestate input (which is a good idea anyway, independent from the laserdisc
stuff).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12803 7512d703-071c-0410-b5c2-d2bdd895b7d1