Improve console rendering speed by caching
In case the openMSX console is visible (and even without typing anything) the
texts in the console were constantly (= each frame) being re-rendered. This
rendering takes time. In case of openGL we also need to upload the rendered
text surface to a openGL texture before it can actually be drawn.

These rendered texts are often exactly the same between successive frames, so
adding a cache should be very effective. That's what this patch does.

I tested by booting a MSX2 to BASIC and opening the openMSX console (without
typing anything). When using the SDLGL-PP renderer, total openMSX CPU usage
dropped from approx 17% to 9%. When using the SDL renderer, the speed gain was
less but still measurable: from approx 26% to 24%.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12804 7512d703-071c-0410-b5c2-d2bdd895b7d1