[1/5] speedup Tcl stuff: replace vector<TclObject*> with vector<TclObject>
Over time openMSX is relying more and more on Tcl. E.g. we now have lots of
Tcl scripts and some of those scripts are running constantly (e.g. the script
that handles the reversebar). This patch series tries to optimize the way we
integrate Tcl into the rest of openMSX.

When the Tcl core calls the openMSX command dispatch callback, we put each
Tcl_Obj* in a more convenient C++ TclObject wrapper. Before this patch we
new'ed (and later delete'ed) all these wrapper objects. Now we store them by
value.

This is faster (less memory allocation/deallocation, better memory locality).
It also makes the code shorter (both the part that creates/destroys these
wrapper objects and the (many more) parts that use them). Not sure why it
wasn't done like this from the start. Possibly because early versions of
TclObject were not yet copyable.

This patch is quite big, but the majority of the changes are simple mechanical
substitutions.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12805 7512d703-071c-0410-b5c2-d2bdd895b7d1