[2/5] speedup Tcl stuff: removed TclObject::owned member
The Tcl_Obj objects (core Tcl stuff) internally use reference counting. When we
put these objects in our own (c++ friendly) TclObject wrapper we had a member
variable 'owned' to indicate whether we created the Tcl_Obj ourself or we
simply wrap a Tcl_Obj we received from the Tcl core. Though using this boolean
'owned' or using the built-in reference counting has roughly the same cost. And
since the reference counting stuff is anyway present we'd better reuse it.

Speed wise this patch is neutral. But it does make TclObject objects smaller
(which is also good for speed). And it makes the previous and some of the
later patches more effective.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12806 7512d703-071c-0410-b5c2-d2bdd895b7d1