[3/5] speedup Tcl stuff: optimized TclObject::unshare()
Like mentioned in the previous patch TclOjects are internally reference
counted. So the same object can potentionally be shared by many owners. This
means that when you want to change the value of a TclObject you first have to
make sure you are the sole owner (the TclObject class all handles this
internally).

Before this patch we used an internal method unshare(). This would in case of
shared ownership duplicate the underlying Tcl_Obj so that later we can safely
modify our private copy. Though in case we want to modify the value it's
useless to first duplicate the old value.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12807 7512d703-071c-0410-b5c2-d2bdd895b7d1