[4/5] speedup Tcl stuff: auto_ptr<TclObject> -> TclObject
TclObject is a lightweight wrapper around a Tcl_Obj (its internal
representation is only two pointers). Also TclObject is internally reference
counted, so making copies of TclObject is quite cheap.

In some places in the code we create a TclObject and pass it (including
ownership) to some other place. We used to do this via auto_ptr<TclObject>, but
because TclObject copies are cheap it's probably better to just pass them by
value. This saves some dynamic memory allocation. Once we switch to c++11, we
could make TclObject movable and then this operation becomes even cheaper.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12808 7512d703-071c-0410-b5c2-d2bdd895b7d1