[5/5] speedup Tcl stuff: use TclObject in AfterCommand
When you implement a Tcl command in c++ (in the context of openMSX) you have
two choices for the execute() method (a subclass needs to override exactly one
of these two):
 - void execute(const vector<TclObject>& tokens, TclObject& result)
 - string execute(const vector<string>& tokens)
The 1st alternative is faster, but for simple stuff the 2nd is easier.

This patch changes the implementation of the 'after' command from the 2nd to
the 1st alternative because the after command is reasonably often used. E.g.
all our Tcl scripts that periodically need to execute stuff (e.g. the
reversebar) use the after command.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12809 7512d703-071c-0410-b5c2-d2bdd895b7d1