[1/3] speedup SectorBasedDisk::readTrack(): cache last result
In the previous release we added support for DMK images and made disk emulation
in general more accurate. This means that the FDC timing and decisions are now
(also) based on the data in the disk track that is located in between the
actual sector data. Also when not using DMK disk images we emulate the FDC
based on the low level track information (so we translate the data in a DSK
disk image into a standard format track layout). The FDC uses this low level
track data quite frequently. So compared to previous releases we now spend a
lot more time in the SectorBasedDisk::readTrack() method.

Luckily when the FDC requests this track data, it very often requests data for
the same track as in the previous request. So instead of always recalculating
this data, we can cache the last result. Of course we have to invalidate this
cache on a disk write.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12810 7512d703-071c-0410-b5c2-d2bdd895b7d1