[3/3] Speedup CRC calculation of large data blocks
In the previous patch we already added a method to calculate the CRC for a
whole block of memory (instead of one byte per call). Though the implementation
of that new method internally still worked byte-per-byte. This patch replaces
the implementation with a more efficient algorithm. See comments in the code
for more details.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12812 7512d703-071c-0410-b5c2-d2bdd895b7d1