Tweak CPUClock::limitReached()   (use '<' instead of '<=')
Before this patch the end of each emulated Z80 instruction executes code
similar to this:

	remaining -= 5; // amount of cycles for this instruction
	if (remaining <= 0) { ... // continue emulating CPU?

(Actually this is spread over a few inlined functions, so you won't directly
see this sequence in the source code). On x86_64 this generates asm code
similar to this:

	mov    0x58(%rbx),%r10d
	sub    $0x5,%r10d
	test   %r10d,%r10d
	mov    %r10d,0x58(%rbx)
	jg     503758
	jmpq   50383c

This patch makes a small change to the C++ code:

	remaining -= 5;
	if (remaining < 0) { ... // '<=' changed to '<'

Recent gcc versions (newer than gcc-4.7.0) now generate this code:

	subl   $0x5,0x58(%rbx)
	jns    503768
	jmpq   503844

So the compiler can now use the flags set by the 'sub' instruction (no need for
a separate 'test' instruction) and because the value of the 'remaining'
variable is now not needed in a register, the sub instruction can directly
operate on memory (no need for load/store instructions).

BTW: at first sight it may seem that this optimization is also possible on the
original C++ code: replace 'jns' with 'jg' in the generated asm code. Though
this has different behavior when the subtraction underflows (underflow shouldn't
happen in this context, but the compiler can't know this). See also this (invalid)
gcc bug report:
	http://gcc.gnu.org/bugzilla/show_bug.cgi?id=53117

This patch makes the '<=' -> '<' substitution and it adjusts the surrounding
code so that the semantics remain the same. In my tests on x86_64 it saves
about 12kB on the total executable size. Depending on how and what exactly you
measure, I saw a speedup of between 0.4% and 1.8%. I didn't test, but I expect
a smaller gain on ARM (because there only a test instruction can be optimized
away; ARM cannot directly operate on memory). On MIPS I expect no difference in
generated code because there it's not the operation (the sub instruction) that
sets flags, instead the branch instruction does a compare against zero.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12813 7512d703-071c-0410-b5c2-d2bdd895b7d1