Optimized testing for host cpu features (MMX, SSE, ...)
Before this patch you had to do this:
	if (HostCPU::getInstance().hasSSE()) { ...
Now you do this:
	if (HostCPU::hasSSE()) { ...

In both cases the hasXXX() methods/functions are inlined and only return the
value of a boolean variable. getInstance() is a static method that properly
initializes the HostCPU object (when not yet done). In the new approach we put
an explicit call to HostCPU::init() early in the main() function.

Let's start with the disadvantage of the new approach: the new interface
potentially allows to call hasXXX() on an uninitialized HostCPU. This was
impossible with the old approach. We mitigate this risk by putting the
HostCPU::init() call early in the main() function. It is still possible there
are global objects with constructors that run before main(). Though if this
happens (I don't think it occurs in the current code), the hasXXX() methods
return false, so this should be safe (OK, from a correctness POV, but not ideal
from a performance POV).

The advantage of the new approach is that we can omit the check for
is-HostCPU-already-initialized from all caller locations. In all (dynamic)
cases but one, this check is just a boolean variable test. (In the
uninitialized case, the code acquires a mutex and then calls the (non-inlined)
HostCPU constructor). So not having to perform this check saves a bit of
runtime performance and quite some code size (I measured about 8kB).

Typically the HostCPU::hasXXX() tests are done in performance critical code and
the cost of this test has been cut in half. But I could only measure a very
small performance gain (0.1%). Possibly because these tests are very well
predictable by the CPUs branch predictor.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12814 7512d703-071c-0410-b5c2-d2bdd895b7d1