Various variable renames, no functional changes
This performs several renames e.g. sectormap -> sectorMap ... There should be
_no_ functional changes introduced by this patch.

I had this patch sitting in a private branch for some time now (as part of some
general endianness cleanup work (not ready yet)). But because we plan to work on
DirAsDSK soon, I thought it would be best to split off the rename part in order
to reduce future merge conflicts.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12817 7512d703-071c-0410-b5c2-d2bdd895b7d1