Various endianness related cleanups
Added two new utility headers: endian.hh and alignof.hh.

In several places in the code we had duplicate helper functions to read or
write 16 or 32 bit values with a specific endianness. Now all those functions
are centralized in one location (and that single implementation is better
optimized).

DirAsDSK and MSXtar had duplicate structs to represent a FAT12 bootsector and
directory entries. Moved those to a central location. Each time a field of such
a struct was accessed, you had to remember to convert it to the correct
endianness. Now there are helper types that allow to specify the endianness
once at the definition of the struct and later the fields can be accessed using
normal variable reads and assignments.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12823 7512d703-071c-0410-b5c2-d2bdd895b7d1