Configurable joystick ports
- In the machine config file it's now possible to specify
    <JoystickPorts>A</JoystickPorts>
  This allows to emulate machines that only have joystick port A, it's also
  possible to specify only port B ("B"), or no ports at all (""). The default
  value ("AB") means the machine has two joystick ports.
- Moved ownership of JoystickPort from MSXPSG to MSXMotherBoard.

Originally this patch started as a refactoring to support Spectravideo machines
in openMSX. One difference between MSX and SVI machines is that in SVI
machines, the joystick ports are partly connected to the PSG and partly to the
PPI chip (in MSX they are only connected to the PSG). This is similar to how in
an MSX machine the cassette port is partly controlled via the PSG and partly
via the PPI. The cassette port situation is implemented in openMSX by moving
ownership of CassettePort to MSXMotherBoard (instead of MSXPSG or MSXPPI).
Similarly, this patch moves ownership of JoystickPort(s) to MSXMotherBoard.

This refactoring also makes it easier to implement configurable joystick ports,
meaning MSX machines that only have joystick port A or B (or none at all).
(Just like there are MSX machines without a cassette port). So this patch is
not limited to Spectravideo support.

I didn't change any machine configurations yet to actually use only 1 joystick
port.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12824 7512d703-071c-0410-b5c2-d2bdd895b7d1