Fixed latent bug in IDEHD::writeBlockComplete()
This code looked buggy (always writes the same sector data in the loop). Though
it turned out that the scenario in which the bug triggers cannot occur in the
current code (the method can currently never be called for more than one sector
at once). Still it feels better to make writeBlockComplete() more robust.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12826 7512d703-071c-0410-b5c2-d2bdd895b7d1