Cleanup SDL keyboard table
This patch (should) have no functional changes. It only turns the table we use
to map SDL key codes into msx keyboard matrix locations into a much more
compact layout. With layout I mainly mean the layout in the source code: with
the new format is easier to see what's going on (e.g. which keys are mapped and
which are not). But the layout of the table in the generated code is also more
compact (half of the original size).

This refactoring is part of a (work-in-progress) SVI patch. There we need to
add a 2nd keyboard table in the code. And with the new layout it is much easier
to verify that the table is correct. But in isolation this refactoring is also
useful, so I split it off from the SVI patch.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12827 7512d703-071c-0410-b5c2-d2bdd895b7d1