Optimized rendering of superimposed modes (video9000, laserdisc)
I removed some useless work from the rendering path. Internally we use the
method FrameSource::getLineInfo() to get both the line-width and the actual
pixel data for that line. In some cases we call this method when we're only
interested in the line-width, and in those cases it's useless to calculate the
pixel data. This patch adds another method getLineWidth() which avoids that
overhead.

Historically we started with two separate methods getLineWidth() and
getLinePtrImpl(). In revision 10629 both were merged into getLineInfo(). At the
time this was a good idea: all existing implementations of getLineWidth() and
getLinePtrImpl() were extremely cheap (basically just return some pointer), so
it was better to have only one virtual function call and sometimes ignore one
of the results than to (most commonly) have to perform two virtual function
calls. Later we added laserdisc and video9000, which introduced the classes
SuperImposedVideoFrame and SuperImposedFrame. These two classes have a
getLineInfo() method that does relatively expensive work (alpha-blend the lines
of two other FrameSources). So if we ignore that result, it's just wasted
calculations.

This class reintroduces the method getLineWidth() but keeps the getLineInfo()
method unchanged. So now there are two possible ways to get the line-width. The
advantage is that the most common case (getting both pixel data and line width)
still requires only one virtual function call and in case you only need the
line width you don't pay for the calculation of the pixel data. The
disadvantage is that both ways to get the line width should return consistent
results, so there is a bit of code duplication.

After this patch I could measure 10-25% higher frame rates when using video9000
(measured with maxframeskip=0, throttle=off).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12833 7512d703-071c-0410-b5c2-d2bdd895b7d1