Tweak CheckedRam (reduce calls to invalidateMemCache())
Normally, when no umr_callback is set, all the checks in CheckedRam are
bypassed, so that there's literally no overhead in using CheckedRam compared to
the normal Ram class. However this bypass relies on the read/writeCacheLine
stuff, and for the upper region in the Z80 address space (0xff00-0xffff) the
cacheline stuff cannot be used (because that region contains the secondary slot
switching registers).

The effect was that writes (and to a lesser degree also reads) to 0xff00-0xffff
were slightly slower than they could be (even if no umr_callback was set).

This patch makes some tweaks to the CheckedRam code so that the overhead for
reads/writes compared to the plain Ram class is smaller, at least for the common
case that the callback doesn't actually trigger.

I actually couldn't measure any speed difference before/after this patch
(difference was within measurement noise). But I did confirm (by stepping
through the code in a debugger) that the most common code path is now simpler,
and I also counted less calls to invalidateMemCache(), so it must be (a tiny
bit) faster.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12834 7512d703-071c-0410-b5c2-d2bdd895b7d1