Fixed crash on config with overlapping memory devices
Before this patch, openMSX crash on hardwareconfigs which contained something
like the following:
    <ROM id="First small ROM">
      <rom>...</rom>
      <mem base="0x4000" size="0x4000"/>
    </ROM>
    <ROM id="Second bigger, overlapping ROM">
      <rom>...</rom>
      <mem base="0x0000" size="0x10000"/>
    </ROM>

The bug was an exception safety problem in the slot registration routine. This
routine splits the registration of a device at 16kB boundaries. So in the
example above the 1st device can be registered successfully. The second device
is split in 4 chunks, the 1st of those chunks is also registered successfully,
but the 2nd chunk fails (because there's an overlap with the 1st device). At
that point an exception is thrown. During stack unwinding the 2nd ROM device is
destroyed. But later that causes a problem because there's still a pointer to
the (now deleted) device dangling in the slot structure. Normally the
destructor of a device unregisters the device from the slot structure, but this
only happens when the registration was completely successful (in this example
it was partially successful, partially failed).

There are several possible fixes. I choose to split the registration in two
phases: first do all checks (on all chunks), and only if all those checks are
OK, do the actual registration.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12841 7512d703-071c-0410-b5c2-d2bdd895b7d1