[1/16] DirAsDSK: Initial cleanups
After we implemented support for DMK, our DirAsDSK support was broken. This is
because we now emulate the WD2793 at a lower level, at the track level instead
of the logical sector level. For example when you now execute a 'write sector'
WD2793 command, the WD2793 code will retrieve the current track content,
partially update it (write new sector data but also update the CRC checksums)
and write that updated buffer back to the disk back-end. This is perfect for
track based back-ends (like DMK disk images). Though for sector based disk
images (like DSK or DirAsDSK) this means that the above scenario will first
read 9 sectors (to be able to construct a buffer containing the current track
content) followed by a write of 9 sectors (when the updated track buffer is
written back to the disk). And this breaks DirAsDSK because the DirAsDSK
implementation relied on a specific order of sector reads and writes.

This patch series will significantly rewrite DirAsDSK to be able to handle the
new situation. The original motivation for this patch series was to fix the bug
described above. Though I believe the new implementation is also more robust in
general (and the code is also much simpler IMHO). The details of the new
approach will be described in more detail in later steps in the series (e.g.
see step 4 for a summary).

--

This commit itself doesn't make any changes yet to the DirAsDSK algorithm. It
only does several cleanups. Originally all these changes were spread over all
the later steps in this series. Though to make reviewing easier I thought it
was better to split them off and group them all in this patch (so that the
later patches can be more focused on one specific thing).

All changes in this patch are relatively minor:
 - Tweak comments, slight code layout changes.
 - Various variable renames (e.g. in this context it's not very clear what
   'filename' means, I renamed it to either 'hostName' or 'msxName').
 - The method getStartCluster(), was not actually needed, I removed it.
 - Moved (small) helper methods to a more logical location in the file (grouped
   together per category). Also converted methods into static functions if
   possible.
 - Minor (always very localized) tweaks to some implementations: e.g. when
   reading/writing a file sector-per-sector from start to end, it's not needed
   to perform a file-seek before every read/write.
 - ...

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12842 7512d703-071c-0410-b5c2-d2bdd895b7d1