[2/16] DirAsDSK: Remove debug()
The DirAsDSK code contains debug-print code. At the end of this patch series,
those prints are not very useful anymore, because the underlying algorithm has
changed so that the printed info is not relevant anymore. To make the following
patches smaller, this patch removes all debug prints in one go.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12843 7512d703-071c-0410-b5c2-d2bdd895b7d1