[3/16] DirAsDSK: Move functions around
Patches that move a (large) function over a large distance are very hard to
read. That's why I split off all code-shuffling that was originally done in the
later patches in a separate patch. So except for moving functions around there
should be NO changes in this patch.

The order of the methods/functions is now roughly: helper functions, functions for
read, functions for write.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12844 7512d703-071c-0410-b5c2-d2bdd895b7d1