[4/16] DirAsDSK: Only sync with host filesystem after a delay
The key idea in this patch series is to introduce a time aspect in the
algorithm that decides when/what/how to sync the host and the (virtual) msx
filesystem. Before this patch that algorithm was based on the order in which
the sectors were read/written (but as explained in part 1, this is broken since
we added DMK support). After this patch series we use a heuristic based on
(emu)time to detect when the msx is performing an 'action'. An action can e.g.
be copy an existing file to another, so it can involve a (large) sequence of
read/write sector commands. The heuristic is that we assume that during such an
action the msx diskdrive is never idle for more than one second (so the action
itself can take longer than 1s, but there cannot be a pause of 1s during the
action).

BTW this duration of 'one second' may still need fine-tuning. I took one
seconds because in that duration still fits the time to to a seek from track 0
to track 79 (though only at the fastest seek speed) plus one full rotation of
the disk plus some margin. If there are disk roms that seek at a lower speed,
we may need to increase this value (or we could dynamically try to detect that
the drive is performing a seek and if not lower the duration).

Ok, so the idea is to perform a sync from the host to the msx (in other words
import/update/remove host files into the virtual disk) right before the start
of an action, but never during an action. During an action we continuously sync
from the msx to the host (so export all changes made by the msx to the host).
In pseudo code:

    readSector() {
        if (more-than-1s-has passed) {
            do-a-full-sync-from-host-to-msx;
        }
        reset-timer;
        memcpy-sector-from-internal-virtual-disk; // do NOT access host filesystem
    }
    writeSector() {
        reset-timer;
        memcpy-given-data-to-virtual-disk;
        perform-the-change-made-by-this-write-on-the-host;
    }

So during an action, the msx has full control over the virtual disk image. When
the msx reads a sector, that sector directly comes from the virtual disk (e.g.
we DONT anymore read data sectors directly from the host files). And when the
msx writes a sector (only) that sector is changed with the exact data given by
the msx.

During an action, the host filesystem follows directly all changes made by the
msx. Often the MSX needs several sector writes to go from one consistent state
in the FAT filesystem to the next. Though we sync at every sector write. That
means that for (short) periods the state of the host filesystem may not be
correct, though at the end of the action it will be.

---

This patch itself does not yet fully implement the above schema. It only
introduces the timer and the sync-to-host at start-of-action part.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12845 7512d703-071c-0410-b5c2-d2bdd895b7d1