[5/16] DirAsDSK: Flush disk back-end caches
Revision 12810 added a cache for the last read track in SectorBasedDisk. In
case of DirAsDSK this cache must be flushed when the filesystem is changed on
the host side. This patch adds a new method checkCache() in the
SectorAccessibleDisk base class, that method is called right before a subclass
tries to use any of the cached data (this could be the track cache in
SectorBasedDisk, but also the sha1sum cache of the whole disk). The default
implementation of this method is empty, but DirAsDSK overrides it and adds a
similar timeout detection algorithm than added in the previous patch for
read/write sector.

--

There now is a single location in the DirAsDSK source code where the content of
the virtual disk is changed by something else than a msx sector write (the sync
from host to msx). Or in other words where the content of the disk is being changed
"behind the MSX' back". That's a good location to signal disk-was-ejected to the
MSX so that the MSX will flush its internal disk sector caches.

I believe the turbor uses such a mechanism (a disk cache flushed with the
disk-eject signal). Of course if you're using a disk-caching program that ignores
this signal it will still go wrong (just like it will go wrong if you use such
a program and swap the physical floppy disks).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12846 7512d703-071c-0410-b5c2-d2bdd895b7d1