[6/16] DirAsDSK: First delete, then update and last add new
When we import changes from the host filesystem into the virtual msx
filesystem, there can be new or deleted files or the content of existing files
can be changed. This patch reorders the steps in such a sync so that we first
free space in the msx filesystem before allocating new space. Or more
specifically, we first look for deleted host files, next for changed files and
only last add totally new files.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12847 7512d703-071c-0410-b5c2-d2bdd895b7d1