[7/16] DirAsDSK: Don't read live data for host files
Before this patch, when the msx read a sector that was mapped to a host file,
then the resulting data was retrieved directly from the content of that host
file. This patch changes it to always read the content from the virtual msx
disk image.

This change is needed to always get consistent results (from the msx point of
view). Suppose the MSX starts reading a file and shortly after that same file
is changed on the host side. Right before the msx starts reading there is a
sync from host->msx (so the status of the file in the msx filesystem is now in
sync with the host file). Suppose the MSX reads the directory and FAT sectors,
so it knows how long the file is and where (in which sectors) it's located on
the disk. Suppose now the file is changed on the host side. And shortly after
the msx starts reading the data sectors that belong to this file. If we would
retrieve the data directly from the host file, there could be an inconsistency
between the file content (the data sectors) and the meta-data (e.g. the
filesize). OTOH if we retrieve the data sectors from the virtual disk image, we
always have a consistent snapshot, though because it's a snapshot it could be
(depending on the duration of the total msx action) slightly outdated compared
to the state of the host file.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12848 7512d703-071c-0410-b5c2-d2bdd895b7d1