[8/16] DirAsDSK: Fixed bug in FAT chain truncation
In case a host file has shrunk in size, and we sync from host->msx, it's
possible we may need to deallocate some clusters from the old FAT-chain. There
was a bug in that behavior.

Bug was introduced by me in revision 7659 (2008/03/15). This patch fixes it. An
earlier patch in this series already needed similar functionality (free a
complete FAT chain in case a host file got deleted). So we can put that in a
helper function and reuse it.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12849 7512d703-071c-0410-b5c2-d2bdd895b7d1