[9/16] DirAsDSK: Slight changes to writeFATSector()
This patch slightly changes how we handle writes to the FAT sectors. We used
to:
 - on FAT1 write: Only buffer data.
 - on FAT2 write: Compare the data from this sector with the previously
     buffered corresponding FAT1 sector and sync host file(s).
Now we do:
 - on FAT1 write: Only buffer data (= same behaviour as before this patch).
 - on FAT2 write: Sync all the buffered FAT1 changes to the host file system
     (so not only for this specific FAT sector). And don't compare with the
     content of the FAT2 sectors but compare with the old status of FAT1 (so
     update based on actual changes in FAT1 instead of differences between FAT1
     and FAT2).

Note that the actual data that is written to FAT2 is ignored, we only use
writes to FAT2 as a trigger.

See also patch 16 in this series, there this FAT sync behavior is simplified
further.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12850 7512d703-071c-0410-b5c2-d2bdd895b7d1